requirements.txt
Flask==2.2.3
Werkzeug==2.2.2
openai==0.27.4
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.2